+++ hardhat-config/migrations.config.js
@@ -2,5 +2,4 @@ module.exports = {
   paths: {
     deploy: 'migrations',
   },
-  duplicateNetworksWithTags: ['qa'],
 };